Move poll components to the correct folder
